requirements.txt
psycopg2-binary==2.9.3
xgboost==1.6.1
scikit-learn==1.1.1
numpy==1.22.4
pandas==1.4.2
Flask==2.1.2
pytest-flask==1.2.0
pytest-mock==3.3.0